+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Charts: Fixed wrong link and modernized method declaration syntax for weekday function.
 - WFST-Tool: Saving objects now works correctly after layer reordering.
 - Map: Fixed wrong zoom and center if camerea parameters are configured for map 3D.
+- extendLayers: Adjusted initial z-index values to ensure correct layer order on map load.
 
 ---